New post, cleanup.
1. A _site directory was added at some point... this is "build product,"
and therefore should not be part of the commit. I probably did this at
some point.

I added a new blog post, with some diagrams regarding our enduring
understandings.

I added the source for those diagrams, although it is in OmniGraffle,
which is closed-source and Mac-only. But, it is OmniGraffle.

Removed some URLs from the headers/templates that were from the original
site source. (Speaking of which, I want to get an "About" or "Acks" page
up in the "Join Us" category, which acknowledges the open tools that
supported this site.)